added start UI & Added all edit profile features

+++ backend/config/cloudinary.js
@@ -4,21 +4,31 @@ import fs from 'fs'
 export const uploadOnCloudinary = async (filePath) => {
 
     cloudinary.config({
-        cloud_name: process.env.CLOUDINARY_API_NAME,
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
         api_key: process.env.CLOUDINARY_API_KEY,
         api_secret: process.env.CLOUDINARY_API_SECRET
     });
 
     try {
-        if(!filePath){
+        if (!filePath) {
             return null;
         }
-        const uploadResult=await cloudinary.uploader.upload(filePath)
-        fs.unlinkSync(filePath)
-        return uploadResult.secure_url
+
+        const uploadResult = await cloudinary.uploader.upload(filePath);
+
+        if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+        }
+
+        return uploadResult.secure_url;
 
     } catch (error) {
-        fs.unlinkSync(filePath)
-        console.log(error)
+        console.error("CLOUDINARY UPLOAD ERROR:", error);
+
+        if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+        }
+
+        return null;
     }
 }
